Corrigindo bug do arquivo vazio quando não encontrava correspondente no dicionario principal.

+++ dicionario-api/controllers/bundle.js
@@ -1,4 +1,4 @@
-﻿/**
+/**
  * Author: Wesnydy Lima Ribeiro
  * Email: wesnydy@lavid.ufpb.br
  */
@@ -13,6 +13,7 @@ var fs = require('fs')
 //    , shortid = require('shortid')
     , error = require('../helpers/error')
     , http = require("http")
+    , request = require('request')
     , settings = require('../config/settings');
 
 exports.download = function (req, res, next) {
@@ -32,14 +33,15 @@ exports.download = function (req, res, next) {
     var bundleFile = path.join(bundlePlatformDir, req.params.id);
 
     ensureDirectoryExistence(bundleFile);
-    var completed = true;
-    var error = false;
     if (!fs.existsSync(bundleFile)) {
-        completed = false;
         var dic = settings.config.mainDic;
         var port = settings.config.dicPort;
         var url = 'http://' + dic + ':' + port + '/' + req.params.version + '/' + req.params.platform + '/' + req.params.id;
-        getBundle(url, bundleFile, function (completed) {
+        downloadBundle(url, bundleFile, function (callbackerror) {
+            if (callbackerror) {
+                res.status(404);
+                return error.notFound('Can\'t find any content for this version and bundle.', next);
+            }
             res.download(bundleFile, function (err) {
                 if (err)
                     return error.notFound('Can\'t find any content for this version and bundle.', next);
@@ -63,15 +65,52 @@ function ensureDirectoryExistence(filePath) {
     fs.mkdirSync(dirname);
 }
 
+const downloadBundle = (url, dest, cb) => {
+   
+    const sendReq = request.get(url);
+
+    // verify response code
+    sendReq.on('response', (response) => {
+        if (response.statusCode !== 200) {
+            return cb('Response status was ' + response.statusCode);
+        }
+        const file = fs.createWriteStream(dest);
+        sendReq.pipe(file);
+
+        // close() is async, call cb after close completes
+        file.on('finish', () => file.close(cb));
+
+        file.on('error', (err) => { // Handle errors
+            fs.unlink(dest); // Delete the file async. (But we don't check the result)
+            return cb(err.message);
+        });
+    });
+
+    // check for request errors
+    sendReq.on('error', (err) => {
+        fs.unlinkSync(dest);
+        return cb(err.message);
+    });
+
+};
+
 var getBundle = function (url, dest, cb) {
     var file = fs.createWriteStream(dest);
     var request = http.get(url, function (response) {
+        if(err)
         response.pipe(file);
         file.on('finish', function () {
+            console.log("Baixou");
             file.close(cb);  // close() is async, call cb after close completes.
         });
     }).on('error', function (err) { // Handle errors
-        fs.unlink(dest); // Delete the file async. (But we don't check the result)
+        fs.unlink(dest, (err) => {
+            if (err) {
+                console.log("failed to delete local image:" + err);
+            } else {
+                console.log('successfully deleted local image');
+            }
+        }); // Delete the file async. (But we don't check the result)
         if (cb) cb(err.message);
     });
 };
